Tweak regexp to catch less false positives when scanning exception output for filename references.
Previous regexp would be fooled by standard NSLog date prefixes (e.g. “2009-04-06 17:11:12.473”). We now require that there is no ‘:’ in the filename and that the filename has an extension of ‘[a-z]+’.

Ideally we’d use access(path, R_OK) — but this will involve calling ‘/usr/bin/find’ (as done in pathForFilename:).